Refactoring + edit trick done
