+++ src/AppBundle/Resources/views/Home/index.html.twig
@@ -21,7 +21,7 @@
                             <div class="caption">
                                 <h3>{{ trick.getName }}</h3>
                                 <p>{{ trick.getDescription|length > 50 ? trick.getDescription|slice(0, 50) ~ '...' : trick.getDescription }}</p>
-                                <p><a href="#" class="btn btn-primary" role="button">Editer</a> <a href="#"
+                                <p><a href="{{ path('trickViewEdit', {'id': trick.getId}) }}" class="btn btn-primary" role="button">Editer</a> <a href="#"
                                                                                                    class="btn btn-default"
                                                                                                    role="button">Supprimer</a>
                                 </p>
