+++ src/AppBundle/Resources/views/Tricks/view_add.html.twig
